% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTypeTC.R
\name{scTypeTC}
\alias{scTypeTC}
\title{single cell sub-typing of epithelial cells in thyroid cancer with glmnet-lasso model}
\usage{
scTypeTC(test_data = test_data, cutoff_chance = 0.5, sam.name = "test")
}
\arguments{
\item{test_data}{normalized gene expression data matrix from Seurat; genes in rows; cell names in columns.}

\item{cutoff_chance}{cutoff on the chance of consistent predictions to report the prediction results, cell below the cutoff is predicted as undefined,default=0.5}

\item{sam.name}{sample name.}
}
\description{
Predict subtype of single thyroid epithelial cell using multinomial-lasso model following the best vote principles.
Cell with low chance of consistent prediction is labeled as undefined.
}
